+++ src/main/java/com/nisovin/shopkeepers/config/Settings.java
@@ -78,7 +78,7 @@ public class Settings extends Config {
 	/*
 	 * Shop Creation (and removal)
 	 */
-	public static ItemData shopCreationItem = new ItemData(Material.VILLAGER_SPAWN_EGG, "&aShopkeeper", null);
+	public static ItemData shopCreationItem = new ItemData(Material.VILLAGER_SPAWN_EGG, c("&aShopkeeper"), null);
 	public static boolean preventShopCreationItemRegularUsage = true;
 	public static boolean deletingPlayerShopReturnsCreationItem = false;